+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Adding RatesAndBenefits Client (https://developers.vtex.com/vtex-developer-docs/reference/benefits)
 
 ## [2.5.3-beta] - 2021-02-08
 ### Fixed